% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogsTutorial-package.R
\docType{package}
\name{dogsTutorial-package}
\alias{dogsTutorial}
\alias{dogsTutorial-package}
\title{dogsTutorial: Functions about dogs}
\description{
This package provides a set of functions that illustrate how dogs truly are man's best friend.
}
\author{
\strong{Maintainer}: Nicola nicola.walker@cefas.co.uk Walker (\href{https://orcid.org/0000-0002-9778-0220}{ORCID})

}
\keyword{internal}
